Add tailwindcss and change file structure
