+++ src/app/store.js
@@ -1,0 +1,5 @@
+import { configureStore } from "@reduxjs/toolkit";
+
+export const store = configureStore({
+  reducer: {},
+});
